refactor: CarsMoveStatusHistory 클래스 dto 패키지로 이동시킨다.
