+++ src/main/java/carracing/view/ResultView.java
@@ -2,7 +2,7 @@ package carracing.view;
 
 import carracing.domain.carracing.Car;
 import carracing.domain.carracing.Cars;
-import carracing.domain.carracing.CarsMoveStatusHistory;
+import carracing.domain.carracing.dto.CarsMoveStatusHistory;
 import carracing.domain.carracing.Position;
 
 import java.util.List;
