+++ src/test/java/carracing/domain/carracing/CarRacingTest.java
@@ -1,5 +1,6 @@
 package carracing.domain.carracing;
 
+import carracing.domain.carracing.dto.CarsMoveStatusHistory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
